+++ src/Readarr.Api.V1/Books/UserRatingUpdateResource.cs
@@ -1,0 +1,7 @@
+namespace Readarr.Api.V1.Books
+{
+    public class UserRatingUpdateResource
+    {
+        public decimal? UserRating { get; set; }
+    }
+}